Fix lift
